requirements.txt
streamlit
openai
pandas
python-docx
fpdf
python-docx
openpyxl